requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
numpy>=1.24.0
matplotlib>=3.7.0
tabulate>=0.9.0

# Retrieval components
pinecone-client>=2.2.1
pinecone
python-terrier
openai
dotenv
pyterrier-doc2query # for doc2query indexing
pyterrier-alpha # for docfusion
openai
google-genai
dotenv

# Utilities
tqdm>=4.65.0
joblib>=1.2.0
python-dotenv>=1.0.0
pytest>=7.3.1

# For AWS compatibility
boto3>=1.26.0

# Extras
spacy
textstat 
pinecone-client 
opensearch-py 
rouge_score 
bert_score